+++ src/main/java/com/provedcode/kudos/controller/KudosController.java
@@ -1,5 +1,6 @@
-package com.provedcode.kudos;
+package com.provedcode.kudos.controller;
 
+import com.provedcode.kudos.service.KudosService;
 import com.provedcode.kudos.model.response.KudosAmount;
 import lombok.AllArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -22,5 +23,11 @@ public class KudosController {
     @PostMapping("/proofs/{proof-id}/kudos")
     void addKudosToProof(@PathVariable("proof-id") long id, Authentication authentication) {
         kudosService.addKudosToProof(id, authentication);
+    }
+
+    @PreAuthorize("hasRole('TALENT')")
+    @DeleteMapping("/proofs/{proof-id}/kudos")
+    void deleteKudosFromProof(@PathVariable("proof-id") long id, Authentication authentication) {
+        kudosService.deleteKudosFromProof(id, authentication);
     }
 }
